refactor rollout to take a std::function instead of template function type

+++ include/mcts/policy.hpp
@@ -13,7 +13,7 @@ namespace policy
     {
         // Random rollout policy
         // n_iter denotes amount of simulated games to play from start state
-        double random_rollout(const chess::position &state, chess::side player_turn, std::mt19937 &generator, short n_iter=10)
+        double random_rollout(const chess::position &state, chess::side player_turn, std::mt19937 &generator, int n_iter=10)
         {
             double accumulated_t{0};
             for (int i = 0; i < n_iter; ++i)
